ci(build.yaml): add GitHub Actions workflow to build and push Docker image to GHCR
Add a new GitHub Actions workflow to automate the process of building
and pushing Docker images to GitHub Container Registry (GHCR). This
workflow triggers on pushes to the main branch and can also be manually
dispatched. It sets up Docker Buildx, logs into GHCR, and builds the
Docker image for the linux/amd64 platform, tagging it with 'latest' and
the current commit SHA.